requirements.txt
anyio==3.5.0
asgiref==3.5.1
certifi==2021.10.8
click==8.1.3
fastapi==0.76.0
h11==0.13.0
idna==3.3
psycopg2-binary==2.9.3
pydantic==1.9.0
sniffio==1.2.0
starlette==0.18.0
typing_extensions==4.2.0
uvicorn==0.17.6
